user location is now correctly tracked
